+++ test/reshape.jl
@@ -1,6 +1,6 @@
 module TestReshape
 
-using Test, DataFrames, Random, Logging, PooledArrays, CategoricalArrays
+using Test, DataFrames, Random, PooledArrays, CategoricalArrays
 const ≅ = isequal
 
 @testset "the output of unstack" begin
@@ -143,46 +143,47 @@ end
     df = DataFrame(id=Union{Int, Missing}[1, 2, 1, 2],
                    id2=Union{Int, Missing}[1, 2, 1, 2],
                    variable=["a", "b", "a", "b"], value=[3, 4, 5, 6])
-    @test_logs (:warn, "Duplicate entries in unstack at row 3 for key 1 and variable a.") unstack(df, :id, :variable, :value)
-    @test_logs (:warn, "Duplicate entries in unstack at row 3 for key (1, 1) and variable a.") unstack(df, :variable, :value)
-    a, b = with_logger(NullLogger()) do
-        unstack(df, :id, :variable, :value), unstack(df, :variable, :value)
-    end
+    @test_throws ArgumentError unstack(df, :id, :variable, :value)
+    @test_throws ArgumentError unstack(df, :variable, :value)
+    a = unstack(df, :id, :variable, :value, allowduplicates=true)
+    b = unstack(df, :variable, :value, allowduplicates=true)
     @test a ≅ DataFrame(id = [1, 2], a = [5, missing], b = [missing, 6])
     @test b ≅ DataFrame(id = [1, 2], id2 = [1, 2], a = [5, missing], b = [missing, 6])
 
     df = DataFrame(id=1:2, variable=["a", "b"], value=3:4)
-    @test_nowarn unstack(df, :id, :variable, :value)
-    @test_nowarn unstack(df, :variable, :value)
     a = unstack(df, :id, :variable, :value)
     b = unstack(df, :variable, :value)
     @test a ≅ b ≅ DataFrame(id = [1, 2], a = [3, missing], b = [missing, 4])
 
     df = DataFrame(variable=["x", "x"], value=[missing, missing], id=[1,1])
-    @test_logs (:warn, "Duplicate entries in unstack at row 2 for key 1 and variable x.") unstack(df, :variable, :value)
-    @test_logs (:warn, "Duplicate entries in unstack at row 2 for key 1 and variable x.") unstack(df, :id, :variable, :value)
+    @test_throws ArgumentError unstack(df, :variable, :value)
+    @test_throws ArgumentError unstack(df, :id, :variable, :value)
+    @test unstack(df, :variable, :value, allowduplicates=true) ≅ DataFrame(id=1, x=missing)
+    @test unstack(df, :id, :variable, :value, allowduplicates=true) ≅ DataFrame(id=1, x=missing)
 end
 
 @testset "missing values in colkey" begin
     df = DataFrame(id=[1, 1, 1, missing, missing, missing, 2, 2, 2],
                    variable=["a", "b", missing, "a", "b", "missing", "a", "b", "missing"],
                    value=[missing, 2.0, 3.0, 4.0, 5.0, missing, 7.0, missing, 9.0])
-    @test_logs (:warn, "Missing value in variable :variable at row 3. Skipping.") unstack(df, :variable, :value)
-    udf = with_logger(NullLogger()) do
-        unstack(df, :variable, :value)
-    end
-    @test propertynames(udf) == [:id, :a, :b, :missing]
+    @test_throws ArgumentError unstack(df, :variable, :value)
+    @test_throws ArgumentError unstack(df, :variable, :value, allowmissing=true)
+    udf = unstack(df, :variable, :value, allowmissing=true, renamecols=x -> coalesce(x, "MISSING"))
+    @test propertynames(udf) == [:id, :a, :b, :missing, :MISSING]
     @test udf[!, :missing] ≅ [missing, 9.0, missing]
+    @test udf[!, :MISSING] ≅ [3.0, missing, missing]
+
     df = DataFrame(id=[1, 1, 1, missing, missing, missing, 2, 2, 2],
                    id2=[1, 1, 1, missing, missing, missing, 2, 2, 2],
                    variable=["a", "b", missing, "a", "b", "missing", "a", "b", "missing"],
                    value=[missing, 2.0, 3.0, 4.0, 5.0, missing, 7.0, missing, 9.0])
-    @test_logs (:warn, "Missing value in variable :variable at row 3. Skipping.") unstack(df, 3, 4)
-    udf = with_logger(NullLogger()) do
-        unstack(df, 3, 4)
-    end
-    @test propertynames(udf) == [:id, :id2, :a, :b, :missing]
+    @test_throws ArgumentError unstack(df, 3, 4)
+    @test_throws ArgumentError unstack(df, 3, 4, allowmissing=true)
+    udf = unstack(df, 3, 4, allowmissing=true, renamecols=x -> coalesce(x, "MISSING"))
+
+    @test propertynames(udf) == [:id, :id2, :a, :b, :missing, :MISSING]
     @test udf[!, :missing] ≅ [missing, 9.0, missing]
+    @test udf[!, :MISSING] ≅ [3.0, missing, missing]
 end
 
 @testset "stack-unstack correctness" begin
@@ -419,23 +420,25 @@ end
     @test_throws ArgumentError flatten(df_bad, [:b, :c])
 end
 
-@testset "test RepeatedVector for categorical" begin
-    v = categorical(["a", "b", "c"], ordered=true)
-    levels!(v, ["b", "c", "a"])
-    rv = DataFrames.RepeatedVector(v, 1, 1)
-    @test isordered(v)
-    @test isordered(categorical(v))
-    @test levels(v) == ["b", "c", "a"]
-    @test levels(categorical(v)) == ["b", "c", "a"]
+# TODO: uncomment these tests when we improve handling of categorical here
 
-    v = categorical(["a", "b", "c"])
-    levels!(v, ["b", "c", "a"])
-    rv = DataFrames.RepeatedVector(v, 1, 1)
-    @test !isordered(v)
-    @test !isordered(categorical(v))
-    @test levels(v) == ["b", "c", "a"]
-    @test levels(categorical(v)) == ["b", "c", "a"]
-end
+# @testset "test RepeatedVector for categorical" begin
+#     v = categorical(["a", "b", "c"], ordered=true)
+#     levels!(v, ["b", "c", "a"])
+#     rv = DataFrames.RepeatedVector(v, 1, 1)
+#     @test isordered(rv)
+#     @test isordered(categorical(rv))
+#     @test levels(rv) == ["b", "c", "a"]
+#     @test levels(categorical(rv)) == ["b", "c", "a"]
+
+#     v = categorical(["a", "b", "c"])
+#     levels!(v, ["b", "c", "a"])
+#     rv = DataFrames.RepeatedVector(v, 1, 1)
+#     @test !isordered(rv)
+#     @test !isordered(categorical(rv))
+#     @test levels(rv) == ["b", "c", "a"]
+#     @test levels(categorical(rv)) == ["b", "c", "a"]
+# end
 
 @testset "stack categorical test" begin
     Random.seed!(1234)
@@ -505,6 +508,71 @@ end
     @test eltype(typeof(sdf2.variable)) === String
     @test eltype(sdf2.value) === Float64
     @test eltype(typeof(sdf2.value)) === Float64
+end
+
+@testset "additional unstack tests" begin
+    df = DataFrame(id=repeat(1:3, inner=3),
+                   id2=repeat(1:3, inner=3),
+                   var=repeat('a':'c', 3),
+                   val=1:9)
+    @test unstack(df, :id, :var, :val) == DataFrame(id=1:3, a=1:3:7, b=2:3:8, c=3:3:9)
+    @test unstack(df, [:id, :id2], :var, :val) == unstack(df, :var, :val) ==
+          DataFrame(id=1:3, id2=1:3, a=1:3:7, b=2:3:8, c=3:3:9)
+
+    Random.seed!(1234)
+    for i in 1:3
+        df = df[Random.shuffle(1:9), :]
+        @test unstack(df, :id, :var, :val) == DataFrame(id=1:3, a=1:3:7, b=2:3:8, c=3:3:9)
+        @test unstack(df, [:id, :id2], :var, :val) == unstack(df, :var, :val) ==
+              DataFrame(id=1:3, id2=1:3, a=1:3:7, b=2:3:8, c=3:3:9)
+    end
+
+    df = DataFrame(id=repeat(1:3, inner=3),
+                   a=repeat(1:3, inner=3),
+                   var=repeat('a':'c', 3),
+                   val=1:9)
+    @test unstack(df, :id, :var, :val) == DataFrame(id=1:3, a=1:3:7, b=2:3:8, c=3:3:9)
+    @test_throws ArgumentError unstack(df, :a, :var, :val)
+    @test_throws ArgumentError unstack(df, [:id, :a], :var, :val)
+
+    df = DataFrame(id=repeat(1:3, inner=3),
+                   id2=repeat(1:3, inner=3),
+                   var=repeat('a':'c', 3),
+                   val=1:9)
+    df[4, 1:2] .= 1
+    @test_throws ArgumentError unstack(df, :id, :var, :val)
+    @test_throws ArgumentError unstack(df, [:id, :id2], :var, :val)
+    @test unstack(df, :id, :var, :val, allowduplicates=true) ≅
+          DataFrame(id=1:3, a=[4, missing, 7], b=2:3:8, c=3:3:9)
+    @test unstack(df, [:id, :id2], :var, :val, allowduplicates=true) ≅
+          DataFrame(id=1:3, id2=1:3, a=[4, missing, 7], b=2:3:8, c=3:3:9)
+
+    df = DataFrame(id=repeat(1:3, inner=3),
+                   id2=repeat(1:3, inner=3),
+                   var=repeat('a':'c', 3),
+                   val=1:9)
+    allowmissing!(df, :var)
+    df.var[4] = missing
+    @test_throws ArgumentError unstack(df, :id, :var, :val)
+    @test_throws ArgumentError unstack(df, [:id, :id2], :var, :val)
+    @test unstack(df, :id, :var, :val, allowmissing=true) ≅
+          DataFrame(id=1:3, a=[1, missing, 7], b=2:3:8, c=3:3:9, missing=[missing, 4, missing])
+    @test unstack(df, [:id, :id2], :var, :val, allowmissing=true) ≅
+          DataFrame(id=1:3, id2=1:3, a=[1, missing, 7], b=2:3:8, c=3:3:9, missing=[missing, 4, missing])
+
+    # test scenario when sorting fails both in grouping and in variable
+    struct A
+        x
+    end
+
+    df = DataFrame(id=repeat(A.([2, 1, 3]), inner=3),
+                   id2=repeat(A.([2, 1, 3]), inner=3),
+                   var=repeat(A.([3, 2, 1]), 3),
+                   val=1:9)
+    @test unstack(df, :id, :var, :val, renamecols=x -> Symbol(:x, x.x)) ==
+          DataFrame(id=A.([2, 1, 3]), x3=1:3:7, x2=2:3:8, x1=3:3:9)
+    @test unstack(df, [:id, :id2], :var, :val, renamecols=x -> Symbol(:x, x.x)) ==
+          DataFrame(id=A.([2, 1, 3]), id2=A.([2, 1, 3]), x3=1:3:7, x2=2:3:8, x1=3:3:9)
 end
 
 end # module